feat: extract types
