+++ src/search/index.ts
@@ -1,5 +1,6 @@
 export * from './cosmoz-dropdown-search';
-export { getSearchInput, getVisibleItems, makeElementsFocusable, handleKeyboardEvent, type KeyboardContext } from './navigation-utils';
-export * from './use-navigation';
+export { getSearchInput, getVisibleItems, makeElementsFocusable, handleKeyboardEvent } from './navigation-utils';
+export { useSearchNavigation, type UseSearchNavigationOptions } from './use-navigation';
 export * from './use-search-filter';
 export * from './use-visible-items';
+export type { Position, KeyboardContext } from './types';
